+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#6](https://github.com/LaxarJS/ax-media-widget/issues/6): Use scope.resources instead of scope.model with LaxarJS patterns resource-handler.
 - [#5](https://github.com/LaxarJS/ax-media-widget/issues/5): Documentation: Fixed highlight resource and layout of section actions
 - [#4](https://github.com/LaxarJS/ax-media-widget/issues/4): Documentation: Fixed layout in section "Integration" and fixed anchor link to this section
 - [#3](https://github.com/LaxarJS/ax-media-widget/issues/3): Documentation: Revised R4.1